+++ src/main/java/com/duktown/global/security/handler/CustomAuthenticationSuccessHandler.java
@@ -4,7 +4,7 @@ import com.duktown.domain.user.dto.UserDto;
 import com.duktown.domain.user.entity.User;
 import com.duktown.domain.user.entity.UserRepository;
 import com.duktown.global.security.service.CustomUserDetails;
-import com.duktown.global.security.service.JwtTokenProvider;
+import com.duktown.global.security.provider.JwtTokenProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
